requirements.txt
requests==2.28.0
requests-cache==0.9.4
requests-oauthlib==1.3.1
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
cattrs==1.10.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.5
django-crispy-forms==1.14.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
frozenlist==1.3.0
idna==3.3
inflect==0.2.5
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
oauthlib==3.2.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-mock==3.7.0
python3-openid==3.2.0
pytz==2022.1
six==1.10.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.9
yarl==1.7.2